+++ .changeset/thin-pens-sparkle.md
@@ -1,0 +1,5 @@
+---
+"curioucity": patch
+---
+
+Separate grpc_discord to guild, thread, message test files
