+++ tests/api/grpc.js
@@ -1,6 +1,8 @@
 import * as urlServices from "./grpc_url.js";
 import * as tagServices from "./grpc_tag.js";
-import * as discordServices from "./grpc_discord.js";
+import * as discordGuildServices from "./grpc_discord_guild.js";
+import * as discordThreadServices from "./grpc_discord_thread.js";
+import * as discordMessageServices from "./grpc_discord_message.js";
 
 export const GRPC_API_HOST = "localhost:8080";
 
@@ -15,19 +17,11 @@ const grpc = () => {
   // tagServices.createTag();
   // tagServices.deleteTag();
   // tagServices.cleanUpTags();
-  discordServices.listDiscordGuild();
-  discordServices.createDiscordGuild();
-  discordServices.deleteDiscordGuild();
-  discordServices.getDiscordGuild();
-  discordServices.cleanupDiscordGuild();
-  // discordServices.listDiscordThread();
-  // discordServices.createDiscordThread();
-  // discordServices.deleteDiscordThread();
-  // discordServices.getDiscordThread();
-  // discordServices.listDiscordMessage();
-  // discordServices.createDiscordMessage();
-  // discordServices.deleteDiscordMessage();
-  // discordServices.getDiscordMessage();
+  discordGuildServices.listDiscordGuild();
+  discordGuildServices.createDiscordGuild();
+  discordGuildServices.deleteDiscordGuild();
+  discordGuildServices.getDiscordGuild();
+  discordGuildServices.cleanupDiscordGuild();
 };
 
 export default grpc;
